src/main/java/com/d1/controller/UserController.java
package com.d1.controller;

import com.d1.domain.User;
import com.d1.repository.UserRepository;
import com.d1.service.UserService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class UserController {

    private final UserService userService;
    private final UserRepository userRepository;

    public UserController(UserService userService, UserRepository userRepository) {
        this.userService = userService;
        this.userRepository = userRepository;
    }

    @GetMapping("/users")
    public List<User> getUsers() {
        return userRepository.findAll();
    }

    @PostMapping("/user/create")
    public void create(@RequestParam String userName, @RequestParam String gameName) {
        userService.create(userName, gameName);
    }
}
